Regroup native classes of encoding module
