+++ src/main/java/org/ballerinalang/stdlib/encoding/nativeimpl/Encode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2019, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -11,22 +11,31 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  */
 
 package org.ballerinalang.stdlib.encoding.nativeimpl;
 
+import org.ballerinalang.jvm.values.ArrayValue;
 import org.ballerinalang.stdlib.encoding.EncodingUtil;
 
 import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
+import java.util.Base64;
 
 /**
- * Extern function to encode URI components.
- * ballerina/encoding:encodeUriComponent
+ * Extern functions of ballerina encoding.
+ *of
+ * @since 0.991.0
  */
-public class EncodeUriComponent {
+public class Encode {
+
+    public static String encodeBase64Url(ArrayValue input) {
+        byte[] encodedValue = Base64.getUrlEncoder().encode(input.getBytes());
+        return new String(encodedValue, StandardCharsets.ISO_8859_1);
+    }
 
     public static Object encodeUriComponent(String url, String charset) {
         try {
